# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Self-Healing Test Automation Harness API
  description: |
    A comprehensive REST API for managing test execution, monitoring results, and configuring self-healing capabilities.
    
    ## Features
    - **Test Orchestration**: Execute and manage multiple types of tests (unit, e2e, performance, security)
    - **Self-Healing**: AI-powered test recovery and adaptation
    - **Unified Reporting**: Comprehensive test results and analytics
    - **Plugin Architecture**: Extensible test engine support
    - **Real-time Monitoring**: Live test execution status and progress tracking
    
    ## Authentication
    Currently, the API operates without authentication for development purposes. In production, API key authentication will be implemented.
  version: 1.0.0
  contact:
    name: Test Automation Harness Support
    email: support@test-automation-harness.com
    url: https://github.com/your-org/test-automation-harness
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.test-automation-harness.com
    description: Production server
security:
  - ApiKeyAuth: []
paths:
...
  /api/v1/tests/execute:
    post:
      summary: Execute a new test
      description: Start execution of a new test with the specified configuration
      tags:
        - Test Execution
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestConfig'
            examples:
              playwright_test:
                summary: Playwright E2E Test
                value:
                  name: "Login Test"
                  description: "Test user login functionality"
                  engine: "playwright"
                  config:
                    url: "https://example.com"
                    timeout: 30000
                    browser: "chromium"
                    headless: true
                  options:
                    timeout: 30000
                    retries: 2
                    parallel: false
                    healing: true
              jest_test:
                summary: Jest Unit Test
                value:
                  name: "User Service Test"
                  description: "Test user service functionality"
                  engine: "jest"
                  config:
                    testFile: "user.service.test.js"
                    timeout: 10000
                  options:
                    timeout: 10000
                    retries: 1
                    parallel: true
                    healing: false
      responses:
        '202':
          description: Test execution started successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                message: "Test execution started successfully"
                data:
                  testId: "test_Login_Test_abc123"
                  status: "accepted"
                  message: "Test execution started"
                  estimatedDuration: 30000
                statusCode: 202
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/tests/{id}/status:
    get:
      summary: Get test execution status
      description: Get the current status of a test execution
      tags:
        - Test Execution
      parameters:
        - name: id
          in: path
          required: true
          description: Test ID
          schema:
            type: string
            example: "test_Login_Test_abc123"
      responses:
        '200':
          description: Test status information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                data:
                  testId: "test_Login_Test_abc123"
                  status: "running"
                  startTime: "2025-01-06T15:30:00.000Z"
                  endTime: null
                  duration: 0
                  progress: 45
                  healingAttempts: 0
                  errors: 0
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/v1/tests/{id}/result:
    get:
      summary: Get complete test result
      description: Get the complete result of a test execution
      tags:
        - Test Execution
      parameters:
        - name: id
          in: path
          required: true
          description: Test ID
          schema:
            type: string
            example: "test_Login_Test_abc123"
        - name: includeArtifacts
          in: query
          description: Include test artifacts
          schema:
            type: boolean
            default: false
        - name: includeHealingAttempts
          in: query
          description: Include healing attempts
          schema:
            type: boolean
            default: true
        - name: includeMetrics
          in: query
          description: Include performance metrics
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Complete test result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                data:
                  testId: "test_Login_Test_abc123"
                  status: "passed"
                  startTime: "2025-01-06T15:30:00.000Z"
                  endTime: "2025-01-06T15:30:30.000Z"
                  duration: 30000
                  engine: "playwright"
                  config: {}
                  results:
                    passed: 5
                    failed: 0
                    skipped: 0
                    total: 5
                  artifacts: {}
                  healingAttempts: []
                  metrics: {}
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/v1/tests/{id}/cancel:
    post:
      summary: Cancel a running test
      description: Cancel the execution of a running test
      tags:
        - Test Execution
      parameters:
        - name: id
          in: path
          required: true
          description: Test ID
          schema:
            type: string
            example: "test_Login_Test_abc123"
      responses:
        '200':
          description: Test cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                data:
                  testId: "test_Login_Test_abc123"
                  status: "cancelled"
                  message: "Test execution cancelled"
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/v1/tests/queue:
    get:
      summary: Get test execution queue status
      description: Get the current status of the test execution queue
      tags:
        - Test Execution
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortParam'
        - $ref: '#/components/parameters/SortByParam'
      responses:
        '200':
          description: Test queue status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                data:
                  queue:
                    - testId: "test_123"
                      name: "Test Name"
                      engine: "playwright"
                      status: "queued"
                      createdAt: "2025-01-06T15:30:00.000Z"
                  running: 2
                  completed: 15
                  failed: 1
                  total: 18
                  pagination:
                    page: 1
                    limit: 10
                    total: 18
                    totalPages: 2
                    hasNext: true
                    hasPrev: false
  /api/v1/tests/engines:
    get:
      summary: Get available test engines
      description: Get list of available test engines
      tags:
        - Test Execution
      parameters:
        - name: testType
          in: query
          description: Filter by test type
          schema:
            type: string
            enum: [unit, e2e, performance, security]
        - name: status
          in: query
          description: Filter by status
          schema:
            type: string
            enum: [enabled, disabled]
        - name: supportsHealing
          in: query
          description: Filter by healing support
          schema:
            type: boolean
      responses:
        '200':
          description: Available test engines
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                data:
                  - name: "playwright"
                    version: "1.40.0"
                    type: "e2e"
                    enabled: true
                    supportsHealing: true
                    capabilities: ["web", "mobile", "api"]
                    config: {}
...
  /api/v1/engines:
    get:
      summary: Get all available test engines
      description: Get list of all available test engines with filtering options
      tags:
        - Engine Management
      parameters:
        - name: testType
          in: query
          description: Filter by test type
          schema:
            type: string
            enum: [unit, e2e, performance, security]
        - name: status
          in: query
          description: Filter by status
          schema:
            type: string
            enum: [enabled, disabled]
        - name: supportsHealing
          in: query
          description: Filter by healing support
          schema:
            type: boolean
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortParam'
        - $ref: '#/components/parameters/SortByParam'
      responses:
        '200':
          description: Available test engines
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                data:
                  - id: "playwright-1.40.0"
                    name: "playwright"
                    version: "1.40.0"
                    type: "e2e"
                    enabled: true
                    supportsHealing: true
                    config:
                      browser: "chromium"
                      headless: true
                      timeout: 30000
                    capabilities: ["web", "mobile", "api"]
                    createdAt: "2025-01-06T15:30:00.000Z"
                    updatedAt: "2025-01-06T15:30:00.000Z"
                pagination:
                  page: 1
                  limit: 10
                  total: 5
                  totalPages: 1
                  hasNext: false
                  hasPrev: false
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (currently optional in development)
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    SortParam:
      name: sort
      in: query
      description: Sort order
      schema:
        type: string
        enum: [asc, desc]
        default: desc
    SortByParam:
      name: sortBy
      in: query
      description: Sort field
      schema:
        type: string
        default: createdAt
  schemas:
    TestConfig:
      type: object
      required:
        - name
        - engine
        - config
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Test name
          example: "Login Test"
        description:
          type: string
          maxLength: 1000
          description: Test description
          example: "Test user login functionality"
        engine:
          type: string
          minLength: 1
          maxLength: 100
          description: Test engine name
          example: "playwright"
        config:
          type: object
          description: Engine-specific configuration
          additionalProperties: true
          example:
            url: "https://example.com"
            timeout: 30000
            browser: "chromium"
            headless: true
        options:
          type: object
          properties:
            timeout:
              type: integer
              minimum: 1000
              maximum: 300000
              default: 30000
              description: Test timeout in milliseconds
            retries:
              type: integer
              minimum: 0
              maximum: 5
              default: 0
              description: Number of retries on failure
            parallel:
              type: boolean
              default: false
              description: Whether to run tests in parallel
            healing:
              type: boolean
              default: true
              description: Whether to enable self-healing
          description: Test execution options
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          description: Success message
          example: "Operation completed successfully"
        data:
          type: object
          description: Response data
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
          example: "2025-01-06T15:30:00.000Z"
        statusCode:
          type: integer
          description: HTTP status code
          example: 200
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            type:
              type: string
              description: Error type
              example: "ValidationError"
            message:
              type: string
              description: Error message
              example: "Request validation failed"
            statusCode:
              type: integer
              description: HTTP status code
              example: 400
            requestId:
              type: string
              description: Request identifier
              example: "req_abc123"
            details:
              type: string
              description: Additional error details
              example: "Field 'name' is required"
            field:
              type: string
              description: Field that caused the error (for validation errors)
              example: "name"
            value:
              type: string
              description: Invalid value (for validation errors)
              example: ""
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2025-01-06T15:30:00.000Z"
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              type: "ValidationError"
              message: "Request validation failed"
              statusCode: 400
              requestId: "req_abc123"
              details: "Field 'name' is required"
              field: "name"
              value: ""
            timestamp: "2025-01-06T15:30:00.000Z"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              type: "NotFoundError"
              message: "Resource not found"
              statusCode: 404
              requestId: "req_abc123"
            timestamp: "2025-01-06T15:30:00.000Z"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              type: "InternalServerError"
              message: "Internal server error"
              statusCode: 500
              requestId: "req_abc123"
            timestamp: "2025-01-06T15:30:00.000Z"
...