# src/api/routes/testExecution.ts
/**
 * Test execution API routes for the Self-Healing Test Automation Harness
 * 
 * This module provides REST API endpoints for executing tests, managing test runs,
 * and monitoring test execution status.
 */

import { Router, Request, Response } from 'express';
import Joi from 'joi';
import { logger } from '../../utils/logger';
import { 
  requestValidationMiddleware, 
  CommonSchemas
} from '../middleware/validation';
import { 
  asyncHandler, 
  createSuccessResponse,
  ApiError,
  NotFoundError,
  ConflictError,
  ValidationError
} from '../middleware/errorHandler';
import { TestResult } from '../../types';

// Simplified interface for API requests
interface ApiTestConfig {
  name: string;
  description?: string;
  engine: string;
  config: Record<string, any>;
  options?: {
    timeout?: number;
    retries?: number;
    parallel?: boolean;
    healing?: boolean;
  };
}

// TODO: Import actual services when they're implemented
// import { TestOrchestrator } from '../../core/TestOrchestrator';
// import { TestEngineFactory } from '../../core/TestEngineFactory';

import { testRuns, testQueue, syncTestRunToResults } from '../storage/shared';

const router = Router();

/**
 * POST /api/v1/tests/execute
 * Execute a new test
 */
router.post('/execute', 
  asyncHandler(async (req: Request, res: Response) => {
    const requestId = (req as any).requestId;
    const testConfig: ApiTestConfig = req.body;

    // Basic validation
    if (!testConfig.name || testConfig.name.trim() === '') {
      throw new ValidationError('Test name is required and cannot be empty', 'name');
    }
    if (!testConfig.engine || testConfig.engine.trim() === '') {
      throw new ValidationError('Test engine is required', 'engine');
    }
    if (!testConfig.config) {
      throw new ValidationError('Test configuration is required', 'config');
    }

    logger.info('Test execution requested', {
      requestId,
      testName: testConfig.name,
      engine: testConfig.engine,
    });

    try {
      // TODO: Replace with actual test orchestration
      // const orchestrator = new TestOrchestrator();
      // const result = await orchestrator.executeTest(testConfig);

      // Demo implementation
      const testId = generateTestId(testConfig.name);
      const result: TestResult = {
        id: testId,
        name: testConfig.name,
        status: 'running',
        startTime: new Date(),
        duration: 0,
        output: '',
        errors: [],
        metrics: {
          memoryUsage: 0,
          cpuUsage: 0,
          networkRequests: 0,
          custom: {},
        },
        healingAttempts: [],
        artifacts: [],
      };

      // Store test run
      testRuns.set(testId, result);
      testQueue.set(testId, testConfig);
      
      // Sync to test results for the results endpoint
      syncTestRunToResults(testId, result);

      // Simulate async test execution
      setTimeout(() => {
        simulateTestExecution(testId, testConfig);
      }, 100);

      res.status(202).json(createSuccessResponse(
        {
          testId,
          status: 'accepted',
          message: 'Test execution started',
          estimatedDuration: testConfig.options?.timeout || 30000,
        },
        'Test execution started successfully',
        202
      ));

    } catch (error) {
      logger.error('Test execution failed', {
        requestId,
        error: (error as Error).message,
        testConfig,
      });

      if (error instanceof ApiError) {
        throw error;
      }

      throw new ApiError(
        'Failed to start test execution',
        500,
        true,
        requestId
      );
    }
  })
);

/**
 * GET /api/v1/tests/:id/status
 * Get test execution status
 */
router.get('/:id/status',
  requestValidationMiddleware({
    params: CommonSchemas.idParam,
  }),
  asyncHandler(async (req: Request, res: Response) => {
    const requestId = (req as any).requestId;
    const testId = req.params['id'] as string;

    logger.debug('Test status requested', {
      requestId,
      testId,
    });

    const testResult = testRuns.get(testId);
    
    if (!testResult) {
      throw new NotFoundError(`Test run with ID ${testId}`, requestId);
    }

    res.json(createSuccessResponse({
      testId,
      status: testResult.status,
      startTime: testResult.startTime,
      endTime: testResult.endTime,
      duration: testResult.duration,
      progress: calculateProgress(testResult),
      healingAttempts: testResult.healingAttempts.length,
      errors: testResult.errors.length,
    }));

  })
);

/**
 * GET /api/v1/tests/:id/result
 * Get complete test result
 */
router.get('/:id/result',
  requestValidationMiddleware({
    params: CommonSchemas.idParam,
    query: Joi.object({
      includeArtifacts: Joi.boolean().default(false),
      includeHealingAttempts: Joi.boolean().default(true),
      includeMetrics: Joi.boolean().default(true),
    }),
  }),
  asyncHandler(async (req: Request, res: Response) => {
    const requestId = (req as any).requestId;
    const testId = req.params['id'] as string;
    const { includeArtifacts, includeHealingAttempts, includeMetrics } = req.query;

    logger.debug('Test result requested', {
      requestId,
      testId,
      includeArtifacts,
      includeHealingAttempts,
      includeMetrics,
    });

    const testResult = testRuns.get(testId);
    
    if (!testResult) {
      throw new NotFoundError(`Test run with ID ${testId}`, requestId);
    }

    // Filter result based on query parameters
    const result = {
      ...testResult,
      ...(includeArtifacts === 'false' && { artifacts: undefined }),
      ...(includeHealingAttempts === 'false' && { healingAttempts: undefined }),
      ...(includeMetrics === 'false' && { metrics: undefined }),
    };

    res.json(createSuccessResponse(result));

  })
);

/**
 * POST /api/v1/tests/:id/cancel
 * Cancel a running test
 */
router.post('/:id/cancel',
  requestValidationMiddleware({
    params: CommonSchemas.idParam,
  }),
  asyncHandler(async (req: Request, res: Response) => {
    const requestId = (req as any).requestId;
    const testId = req.params['id'] as string;

    logger.info('Test cancellation requested', {
      requestId,
      testId,
    });

    const testResult = testRuns.get(testId);
    
    if (!testResult) {
      throw new NotFoundError(`Test run with ID ${testId}`, requestId);
    }

    if (testResult.status !== 'running') {
      throw new ConflictError(
        `Cannot cancel test with status: ${testResult.status}`,
        requestId
      );
    }

    // TODO: Implement actual test cancellation
    // const orchestrator = new TestOrchestrator();
    // await orchestrator.cancelTest(testId);

    // Demo implementation
    testResult.status = 'failed';
    testResult.endTime = new Date();
    testResult.duration = testResult.endTime.getTime() - testResult.startTime.getTime();
    testResult.output += '\n[INFO] Test execution cancelled by user';

    res.json(createSuccessResponse({
      testId,
      status: 'cancelled',
      message: 'Test execution cancelled successfully',
    }));

  })
);

/**
 * GET /api/v1/tests/queue
 * Get test execution queue status
 */
router.get('/queue',
  asyncHandler(async (req: Request, res: Response) => {
    const requestId = (req as any).requestId;
    const page = parseInt(req.query['page'] as string || '1', 10);
    const limit = parseInt(req.query['limit'] as string || '10', 10);
    const sort = req.query['sort'] as string || 'desc';
    const sortBy = req.query['sortBy'] as string || 'createdAt';

    logger.debug('Test queue status requested', {
      requestId,
      page,
      limit,
      sort,
      sortBy,
    });

    // TODO: Implement actual queue management
    const queueItems = Array.from(testQueue.entries()).map(([id, config]) => ({
      testId: id,
      name: config.name,
      engine: config.engine,
      queuedAt: new Date(), // TODO: Track actual queue time
      priority: 1, // TODO: Implement priority system
    }));

    const total = queueItems.length;
    const startIndex = (page - 1) * limit;
    const endIndex = startIndex + limit;
    const items = queueItems.slice(startIndex, endIndex);

    res.json(createSuccessResponse({
      items,
      pagination: {
        page,
        limit,
        total,
        totalPages: Math.ceil(total / limit),
        hasNext: endIndex < total,
        hasPrev: page > 1,
      },
    }));

  })
);

/**
 * GET /api/v1/tests/engines
 * Get available test engines
 */
router.get('/engines',
  asyncHandler(async (req: Request, res: Response) => {
    const requestId = (req as any).requestId;

    logger.debug('Available engines requested', { requestId });

    // TODO: Get from actual engine registry
    const engines = [
      {
        name: 'playwright',
        version: '1.55.0',
        testType: 'e2e',
        supportsHealing: true,
        status: 'available',
        capabilities: ['web', 'mobile', 'api'],
      },
      {
        name: 'jest',
        version: '30.1.3',
        testType: 'unit',
        supportsHealing: false,
        status: 'available',
        capabilities: ['unit', 'integration'],
      },
      {
        name: 'k6',
        version: '0.50.0',
        testType: 'performance',
        supportsHealing: false,
        status: 'available',
        capabilities: ['load', 'stress', 'spike'],
      },
    ];

    res.json(createSuccessResponse({
      engines,
      total: engines.length,
    }));

  })
);

/**
 * POST /api/v1/tests/batch
 * Execute multiple tests in batch
 */
router.post('/batch',
  requestValidationMiddleware({
    body: Joi.object({
      tests: Joi.array().items(CommonSchemas.testExecution).min(1).max(10).required(),
      options: Joi.object({
        parallel: Joi.boolean().default(false),
        stopOnFailure: Joi.boolean().default(false),
        timeout: Joi.number().integer().min(1000).max(600000).default(300000),
      }).optional(),
    }),
  }),
  asyncHandler(async (req: Request, res: Response) => {
    const requestId = (req as any).requestId;
    const { tests, options = {} } = req.body;

    logger.info('Batch test execution requested', {
      requestId,
      testCount: tests.length,
      parallel: options.parallel,
    });

    // TODO: Implement actual batch execution
    const batchId = generateBatchId();
    const testIds: string[] = [];

    for (const testConfig of tests) {
      const testId = generateTestId(testConfig.name);
      testIds.push(testId);

      const result: TestResult = {
        id: testId,
        name: testConfig.name,
        status: 'running',
        startTime: new Date(),
        duration: 0,
        output: '',
        errors: [],
        metrics: {
          memoryUsage: 0,
          cpuUsage: 0,
          networkRequests: 0,
          custom: {},
        },
        healingAttempts: [],
        artifacts: [],
      };

      testRuns.set(testId, result);
      testQueue.set(testId, testConfig);
    }

    res.status(202).json(createSuccessResponse({
      batchId,
      testIds,
      status: 'accepted',
      message: 'Batch test execution started',
      totalTests: tests.length,
      options,
    }, 'Batch test execution started successfully', 202));

  })
);

// Helper functions

/**
 * Generate a unique test ID
 */
function generateTestId(testName: string): string {
  const timestamp = Date.now().toString(36);
  const random = Math.random().toString(36).substring(2, 8);
  return `test_${testName.replace(/[^a-zA-Z0-9]/g, '_')}_${timestamp}_${random}`;
}

/**
 * Generate a unique batch ID
 */
function generateBatchId(): string {
  const timestamp = Date.now().toString(36);
  const random = Math.random().toString(36).substring(2, 8);
  return `batch_${timestamp}_${random}`;
}

/**
 * Calculate test execution progress
 */
function calculateProgress(testResult: TestResult): number {
  if (testResult.status === 'passed' || testResult.status === 'failed') {
    return 100;
  }
  
  if (testResult.status === 'running') {
    // Simple progress calculation based on time elapsed
    const elapsed = Date.now() - testResult.startTime.getTime();
    const estimated = 30000; // 30 seconds default
    return Math.min(Math.round((elapsed / estimated) * 100), 95);
  }
  
  return 0;
}

/**
 * Simulate test execution (demo purposes)
 */
function simulateTestExecution(testId: string, config: ApiTestConfig): void {
  const testResult = testRuns.get(testId);
  if (!testResult) return;

  // Simulate test execution time
  const executionTime = Math.random() * 10000 + 2000; // 2-12 seconds
  
  setTimeout(() => {
    if (testResult.status === 'running') {
      // Simulate success/failure
      const success = Math.random() > 0.3; // 70% success rate
      
      testResult.status = success ? 'passed' : 'failed';
      testResult.endTime = new Date();
      testResult.duration = testResult.endTime.getTime() - testResult.startTime.getTime();
      
      if (success) {
        testResult.output = `Test ${config.name} completed successfully`;
      } else {
        testResult.errors.push({
          message: 'Simulated test failure',
          stack: 'Error: Simulated failure for demo purposes',
          type: 'unknown',
          timestamp: new Date(),
          context: {},
        });
        testResult.output = `Test ${config.name} failed`;
      }

      logger.info('Simulated test execution completed', {
        testId,
        status: testResult.status,
        duration: testResult.duration,
      });
    }
  }, executionTime);
}

export default router;